[refactor] TariffSummaryCard: apply WidgetCard layout and change to take list of Tariffs
